[BUGFIX]: Disable image processing for unsupported file types
Some CTypes (e.g. uploads) have the option to enable thumbnails. This works for most file types for images and PDFs,
but not for Microsoft-related file types (e.g. .docx). When using this extension on a site that has content elements
(e.g file links with enabled thumbnails)  with file types like mentioned above, an exception will occur, as there
is no image to be processed.

Simply disabling deferred image processing for this unsupported file types should fix this.